test: split focused coordinator test group
