+++ tests/test_coordinator.py
@@ -1,7 +1,6 @@
 """Tests for ThesslaGreenModbusCoordinator - HA 2025.7.1+ & pymodbus 3.5+ Compatible."""
 
 import asyncio
-import logging
 import types
 from types import SimpleNamespace
 from unittest.mock import AsyncMock, MagicMock, patch
@@ -101,45 +100,6 @@ def test_coordinator_clamps_effective_batch():
     )
     assert coord.effective_batch == MAX_BATCH_REGISTERS
     assert coord.max_registers_per_request == MAX_BATCH_REGISTERS
-
-
-@pytest.mark.asyncio
-async def test_read_holding_registers_none_client(coordinator, caplog):
-    """Return empty data when no Modbus client is present."""
-    coordinator.client = None
-    coordinator._register_groups = {"holding_registers": [(0, 1)]}
-
-    with caplog.at_level(logging.DEBUG):
-        result = await coordinator._read_holding_registers_optimized()
-
-    assert result == {}
-    assert "Modbus client is not connected" in caplog.text
-
-
-@pytest.mark.asyncio
-async def test_read_holding_registers_cancelled_error(coordinator, caplog):
-    """Propagate cancellation without logging noise."""
-    coordinator.client = MagicMock()
-    coordinator._register_groups = {"holding_registers": [(0, 1)]}
-    coordinator._call_modbus = AsyncMock(side_effect=asyncio.CancelledError)
-
-    with caplog.at_level(logging.ERROR), pytest.raises(asyncio.CancelledError):
-        await coordinator._read_holding_registers_optimized()
-    assert caplog.text == ""
-
-
-@pytest.mark.asyncio
-async def test_read_input_registers_reconnect_on_error(coordinator):
-    """Ensure disconnect is triggered after Modbus errors."""
-    coordinator.client = MagicMock()
-    coordinator.client.connected = True
-    coordinator._register_groups = {"input_registers": [(0, 1)]}
-    coordinator._call_modbus = AsyncMock(side_effect=ConnectionException("boom"))
-    coordinator._disconnect = AsyncMock()
-
-    await coordinator._read_input_registers_optimized()
-
-    coordinator._disconnect.assert_called()
 
 
 @pytest.mark.asyncio
